+++ lib/View.php
@@ -11,7 +11,7 @@ namespace Wireframe;
  * @property ViewPlaceholders|null $placeholders ViewPlaceholders object.
  * @property Partials|null $partials Object containing partial paths.
  *
- * @version 0.6.2
+ * @version 0.7.0
  * @author Teppo Koivula <teppo@wireframe-framework.com>
  * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
@@ -121,10 +121,18 @@ class View extends \ProcessWire\TemplateFile {
     /**
      * Setter method for the Controller class
      *
-     * @param Controller|null Controller instance or null
+     * @param Controller|string|null Controller instance, Controller class name, or null
      * @return View Self-reference
+     *
+     * @throws Exception if $controller argument is of unrecognized type.
      */
-    public function setController(?Controller $controller): View {
+    public function setController($controller = null): View {
+        if (!\is_null($controller) && !\is_string($controller) && !$controller instanceof Controller) {
+            throw new \Exception('Controller is of unexpected type, please provide a Controller instance, Controller class name (string), or null');
+        }
+        if (\is_string($controller)) {
+            $controller = $this->wire('modules')->get('Wireframe')->getController($this->getPage(), $controller);
+        }
         $this->setViewData('controller', $controller);
         return $this;
     }
@@ -132,10 +140,13 @@ class View extends \ProcessWire\TemplateFile {
     /**
      * Getter method for the Controller class
      *
+     * Note: locally stored Controller reference is intentionally given precedence over Page Controller. This is also
+     * in line with how overriding Page object view and layout behave (no effect on already instantiated View object).
+     *
      * @return Controller|null Controller instance or null
      */
     public function getController(): ?Controller {
-        return $this->getPage()->__wireframe_controller ?: $this->getViewData('controller');
+        return $this->getViewData('controller') ?: $this->getPage()->__wireframe_controller;
     }
 
     /**